PAY-3511 - Test cases fix.
